# shared/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HHSearchProfileInput": {
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "query": {
            "$ref": "#/components/schemas/HHSearchQuery"
          }
        },
        "required": [
          "name",
          "query"
        ],
        "title": "HHSearchProfileInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProfileListResponse": {
        "properties": {
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ProfileListResponse",
        "type": "object"
      },
      "ProfileResponse": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "meta": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-VacancyPilot-Client",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "VacancyPilot Ops Companion",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/hh/auth/disconnect": {
      "post": {
        "operationId": "oauth_disconnect_api_v1_hh_auth_disconnect_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Oauth Disconnect Api V1 Hh Auth Disconnect Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Oauth Disconnect",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/auth/start": {
      "post": {
        "operationId": "oauth_start_api_v1_hh_auth_start_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Oauth Start Api V1 Hh Auth Start Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Oauth Start",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/capabilities": {
      "get": {
        "operationId": "hh_capabilities_api_v1_hh_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Hh Capabilities Api V1 Hh Capabilities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Hh Capabilities",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/search-profiles": {
      "get": {
        "operationId": "list_profiles_api_v1_hh_search_profiles_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Profiles",
        "tags": [
          "hh"
        ]
      },
      "post": {
        "operationId": "create_profile_api_v1_hh_search_profiles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HHSearchProfileInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Profile",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/search-profiles/{profile_id}": {
      "patch": {
        "operationId": "update_profile_api_v1_hh_search_profiles__profile_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HHSearchProfilePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update Profile",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/search-profiles/{profile_id}/preview": {
      "post": {
        "description": "Read-only HH search-size preview; it never touches the database.",
        "operationId": "preview_profile_api_v1_hh_search_profiles__profile_id__preview_post",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__api__hh__PreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Preview Profile",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/sync/applicant": {
      "post": {
        "description": "Read-only applicant projection; no HH mutation and no raw payload storage.",
        "operationId": "sync_applicant_api_v1_hh_sync_applicant_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Sync Applicant Api V1 Hh Sync Applicant Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Sync Applicant",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/sync/negotiations": {
      "post": {
        "description": "Read-only applicant projection; no HH mutation and no raw payload storage.",
        "operationId": "sync_applicant_api_v1_hh_sync_negotiations_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Sync Applicant Api V1 Hh Sync Negotiations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Sync Applicant",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/sync/resumes": {
      "post": {
        "description": "Read-only applicant projection; no HH mutation and no raw payload storage.",
        "operationId": "sync_applicant_api_v1_hh_sync_resumes_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Sync Applicant Api V1 Hh Sync Resumes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Sync Applicant",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/hh/sync/vacancies": {
      "post": {
        "operationId": "sync_vacancies_api_v1_hh_sync_vacancies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VacancySyncRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Sync Vacancies",
        "tags": [
          "hh"
        ]
      }
    },
    "/api/v1/integrations/hh/status": {
      "get": {
        "operationId": "hh_status_api_v1_integrations_hh_status_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Hh Status Api V1 Integrations Hh Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Hh Status",
        "tags": [
          "hh"
        ]
      }
    },
...
  }
}